requirements.txt
transformers>=4.30.0
torch>=2.0.0
datasets>=2.12.0
numpy>=1.24.0
scikit-learn>=1.2.0
pyarrow>=12.0.0
huggingface-hub>=0.16.0
pandas>=2.0.0
pyyaml>=6.0.0
jupyter>=1.0.0
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0 